Safari-611.1.10/LayoutTests/platform/mac/media/encrypted-media/fps-multiple-pssh-expected.txt

RUN(promise = navigator.requestMediaKeySystemAccess("com.apple.fps", capabilities))
Promise resolved OK
RUN(promise = access.createMediaKeys())
Promise resolved OK
RUN(keys.setServerCertificate(serverCertificate))
RUN(session = keys.createSession())
EXPECTED (session != 'null') OK
RUN(promise = session.generateRequest("cenc", initData.buffer))
Promise resolved OK
END OF TEST
